chore: improve readability

+++ cmd/argo-compare/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"github.com/romana/rlog"
 	m "github.com/shini4i/argo-compare/internal/models"
 	"os"
@@ -10,43 +11,54 @@ import (
 type execContext = func(name string, arg ...string) *exec.Cmd
 
 func processFiles(fileName string, fileType string, application m.Application) {
-	rlog.Infof("Processing %s changed files", fileType)
+	rlog.Debugf("Processing %s changed files", fileType)
 
 	app := Application{File: fileName, Type: fileType, App: application}
 	if fileType == "src" {
 		app.parse()
 	}
-	
+
 	app.writeValuesYaml()
 	app.collectHelmChart()
 	app.extractChart()
 	app.renderTemplate()
 }
 
+func compareFiles() {
+	comparer := Compare{}
+	comparer.findFiles()
+	comparer.printCompareResults()
+}
+
 func main() {
-	if _, err := os.Stat("tmp/"); os.IsNotExist(err) {
-		err := os.Mkdir("tmp/", 0755)
-		if err != nil {
-			rlog.Criticalf(err.Error())
-		}
+	repo := GitRepo{}
+
+	changedFiles := repo.getChangedFiles(exec.Command)
+
+	if len(changedFiles) == 0 {
+		fmt.Println("No changed Application files found. Exiting...")
+		os.Exit(0)
 	}
 
-	defer func() {
+	for _, file := range changedFiles {
+		fmt.Println("Processing changed application: ", file)
+		fmt.Println()
+
+		if _, err := os.Stat("tmp/"); os.IsNotExist(err) {
+			err := os.Mkdir("tmp/", 0755)
+			if err != nil {
+				rlog.Criticalf(err.Error())
+			}
+		}
+
+		processFiles(file, "src", m.Application{})
+		app := repo.getChangedFileContent("main", file, exec.Command)
+		processFiles(file, "dst", app)
+		compareFiles()
+
 		err := os.RemoveAll("tmp/")
 		if err != nil {
 			rlog.Criticalf(err.Error())
 		}
-	}()
-
-	repo := GitRepo{}
-
-	for _, file := range repo.getChangedFiles(exec.Command) {
-		processFiles(file, "src", m.Application{})
-		app := repo.getChangedFileContent("main", file, exec.Command)
-		processFiles(file, "dst", app)
 	}
-
-	comparer := Compare{}
-	comparer.findFiles()
-	comparer.printCompareResults()
 }
